fixed req number - auto generation

+++ client/net.h
@@ -36,8 +36,8 @@ struct data_object{
 
 	uint32_t          node_nr_erases;
 	float             local_log_utilization;
-	float			  request_number;
-	float			  response_time;
+	uint64_t			  request_number;
+	uint32_t			  response_time;
 };
 
 bool operator<(const data_object& lhs, const data_object& rhs);
